+++ test/compile/adapter/onerror.js
@@ -2,10 +2,6 @@ const onerror = require('../../../lib/compile/adapter/onerror');
 
 
 module.exports = {
-    before: () => {
-        console.log('#compile/adapter/onerror');
-    },
-
     'onerror': {
         'basic': () => {
             onerror({});