+++ test/compile/index.js
@@ -10,7 +10,6 @@ let render, data, result;
 module.exports = {
 
     before: () => {
-        console.log('#compile/index');
         defaults.onerror = () => {
             return () => '{Template Error}';
         };
@@ -661,7 +660,9 @@ module.exports = {
     'errors': {
         'RuntimeError': {
             'error': () => {
-                const render = compile('<%=a.b.c%>');
+                const render = compile('<%=a.b.c%>', {
+                    bail: false
+                });
                 assert.deepEqual('{Template Error}', render({}));
             },
 
@@ -687,10 +688,14 @@ module.exports = {
 
             'error': () => {
                 let render;
-                render = compile('<%=a b c%>');
+                render = compile('<%=a b c%>', {
+                    bail: false
+                });
                 assert.deepEqual('{Template Error}', render({}));
 
-                render = compile('{{a b c}}');
+                render = compile('{{a b c}}', {
+                    bail: false
+                });
                 assert.deepEqual('{Template Error}', render({}));
             },
 
@@ -732,7 +737,8 @@ module.exports = {
 
             'template not found': () => {
                 const render = compile({
-                    filename: '/404.html'
+                    filename: '/404.html',
+                    bail: false
                 });
                 assert.deepEqual('{Template Error}', render({}));
             },