+++ webpack.config.js
@@ -36,6 +36,7 @@ module.exports = {
     devtool: 'source-map',
     plugins: [
         new webpack.BannerPlugin(`art-template@${version} for browser | https://github.com/aui/art-template`),
+        new webpack.optimize.ModuleConcatenationPlugin(),
         process.env.NODE_ENV === 'production' ? new webpack.optimize.UglifyJsPlugin({
             compress: {
                 warnings: false,